// src/app/model/training.ts
export interface Training {
    id: number;
    title: string;
    description: string;
    category: string; // Assuming category is a string field in Django
    date_start: string; // Assuming date_start is a string field in Django
    date_end: string; // Assuming date_end is a string field in Django
  }
  